reformatted code and change name of test classes

+++ src/main/java/dk/cphbusiness/dat/carporteksamensproject/model/dtos/BillOfMaterialLineItemDTO.java
@@ -18,6 +18,6 @@ public record BillOfMaterialLineItemDTO(BillOfMaterialLineItem lineItem,
     }
 
     public double getPrice() {
-        return (double) lineItem.getAmount() * product.getPrice();
+        return lineItem.getAmount() * product.getPrice();
     }
 }
